refactor : 마이페이지 프로필 이미지 영역 컴포넌트 분리
